chore: move tests to main test file
